Added and then removed cancer

+++ dungeon-game-core/src/com/dungeon/game/item/Bow.java
@@ -73,7 +73,7 @@ public class Bow extends Ranged {
 	}
 
 	@Override
-	public int[] getPos(boolean mousedown, boolean mousepressed) {
+	public float[] getPos(boolean mousedown, boolean mousepressed) {
 		int distance=30;
 		int polarAngle= 10;
 		int angle=0;
@@ -119,7 +119,7 @@ public class Bow extends Ranged {
 			}
 			break;
 		}
-		return new int[]{distance,polarAngle,angle};
+		return new float[]{distance,polarAngle,angle};
 	}
 
 }
